CONTINUE: challenge 1 completed and small part leftover in challenge 2.
