+++ d30_excep_json/main_dupli.py
@@ -1,0 +1,16 @@
+
+fruits = ["Apple", "Pear", "Orange"]
+
+#  TODO: Catch the exception and make sure the code runs without crashing.
+
+
+def make_pie(index):
+    try:
+        fruit = fruits[index]
+    except IndexError:
+        print("Index out of range")
+    else:
+        return print(fruit + " pie")
+
+
+make_pie(2)
